requirements/development.txt
-r base.txt
mock==1.0.1
coverage==3.6
django-debug-toolbar==1.0.1
freezegun==0.1.15
